python/requirements.txt
langchain
langchain-anthropic
langchain-openai
langchain-core
langchain-community
pyyaml
python-dotenv